requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langchain-google-genai==2.0.0
langchain-core==0.3.0
pydantic==2.9.0
python-dotenv==1.0.1

